+++ benchmark/custom_emitter_benchmark.dart
@@ -1,8 +1,9 @@
 // ignore_for_file: unused_local_variable
 import 'package:benchmark_runner/benchmark_runner.dart';
 
-class CustomEmitter extends ColorPrintEmitter {
-  void emitMean({required Score score}) {
+class CustomEmitter implements ScoreEmitter {
+  @override
+  void emit({required description, required Score score}) {
     print('Mean               Standard Deviation');
     print('${score.stats.mean}  ${score.stats.stdDev}');
   }
@@ -14,9 +15,6 @@ void main(List<String> args) {
     () {
       var list = <int>[for (var i = 0; i < 1000; ++i) i];
     },
-    emitter: CustomEmitter(),
-    report: (instance, emitter) => emitter.emitMean(
-      score: instance.score(),
-    ),
+    scoreEmitter: CustomEmitter(),
   );
 }