+++ benchmark/example_async_benchmark.dart
@@ -15,7 +15,7 @@ void main(List<String> args) async {
 
     await asyncBenchmark('5ms', () async {
       await later<int>(27, Duration(milliseconds: 5));
-    }, report: asyncReportMean);
+    }, scoreEmitter: MeanEmitter());
   });
 
   group('Set', () async {