+++ benchmark/example_benchmark.dart
@@ -21,6 +21,6 @@ void main(List<String> args) {
 
     benchmark('construct', () {
       var list = <int>[for (var i = 0; i < 1000; ++i) i];
-    }, report: reportMean);
+    }, scoreEmitter: MeanEmitter());
   });
 }